Moved the AB test helpers into a subfolder.
