+++ assets/components/payPalContributionButton/payPalContributionButton.jsx
@@ -12,7 +12,7 @@ import { generateClassName } from 'helpers/utilities';
 
 import type { IsoCountry } from 'helpers/internationalisation/country';
 import type { ReferrerAcquisitionData } from 'helpers/tracking/acquisitions';
-import type { Participations } from 'helpers/abtest';
+import type { Participations } from 'helpers/abTests/abtest';
 
 
 // ---- Types ----- //
